# docs/openapi.yaml
openapi: 3.0.3
info:
  title: AI Diet Assistant API
  description: |
    RESTful API for AI-powered diet planning and nutrition management system.
    
    ## Features
    - JWT-based authentication
    - Food inventory management
    - Meal tracking and recording
    - AI-powered meal plan generation
    - Nutrition analysis and statistics
    - Conversation with AI assistant
    
    ## Authentication
    Most endpoints require JWT Bearer token authentication.
    Include the token in the Authorization header: `Bearer {token}`
    
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:9090/api/v1
    description: Development server
  - url: https://api.yourdomain.com/api/v1
    description: Production server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login endpoint
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        code:
          type: integer
          example: 0
        message:
          type: string
          example: "success"
        timestamp:
          type: integer
          format: int64
          example: 1699999999
...
    NutritionData:
      type: object
      properties:
        protein:
          type: number
          format: float
          example: 31.0
        carbs:
          type: number
          format: float
          example: 0.0
        fat:
          type: number
          format: float
          example: 3.6
        fiber:
          type: number
          format: float
          example: 0.0
        calories:
          type: number
          format: float
          example: 165.0
...
    MealFood:
      type: object
      properties:
        food_id:
          type: integer
          format: int64
          example: 1
        name:
          type: string
          example: "Chicken Breast"
        amount:
          type: number
          format: float
          example: 200
        unit:
          type: string
          example: "g"
...
    Plan:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        user_id:
          type: integer
          format: int64
          example: 1
        plan_date:
          type: string
          format: date
          example: "2024-01-02"
        meal_type:
          type: string
          enum: [breakfast, lunch, dinner, snack]
          example: "breakfast"
        foods:
          type: array
          items:
            $ref: '#/components/schemas/MealFood'
        nutrition:
          $ref: '#/components/schemas/NutritionData'
        status:
          type: string
          enum: [pending, completed, skipped]
          example: "pending"
        ai_reasoning:
          type: string
          example: "High protein breakfast to start your day"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
paths:
...
  /plans/generate:
    post:
      tags:
        - Plans
      summary: Generate meal plans
      description: Generate AI-powered meal plans for future days
      operationId: generatePlans
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                days:
                  type: integer
                  default: 2
                  example: 2
                preferences:
                  type: string
                  example: "I prefer low-carb meals"
      responses:
        '200':
          description: Plans generated successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Plan'
...